Fixed a problem with some external URLs
There are some download links that do not end with a file extension.
Therefore, the proxy tries to open them to look for additionnal URLs
but PyQuery fails to parse the HTML code because it is a gzipped
tarball or something of the kind.

Example: docutils http://pypi.python.org/simple/docutils/

    http://sourceforge.net/projects/docutils/files/docutils/0.8.1/docutils-0.8.1.tar.gz/download
    http://sourceforge.net/projects/docutils/files/docutils/0.9.1/docutils-0.9.1.tar.gz/download